refactor(inferno-server): Adjust imports and types
