+++ packages/inferno-server/inferno-server.d.ts
@@ -1,0 +1,9 @@
+declare module 'inferno-server' {
+  export function renderToString(input: any): string;
+  export function renderToStaticMarkup(input: any): string;
+
+  export default {
+    renderToString,
+    renderToStaticMarkup
+  };
+}
