+++ packages/inferno-server/inferno-server.js
@@ -1,2 +1,2 @@
-module.exports = require('inferno/dist/inferno-server.node');
+module.exports = require('./dist/inferno-server.node');
 module.exports.default = module.exports;
